server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
ffmpeg-python==0.2.0
librosa==0.11.0
numpy==1.26.4
python-multipart==0.0.20
